fix: remove GitHub Action problem matchers
